refactor(smtp_client): split mod.rs (918 LOC) + purge 148 LOC dead code
Split fonctionnel du client SMTP monolithique:
- body_utils.rs (154 LOC): normalize_crlf, strip_tags_simple, body_looks_like_html,
  ensure_html_document, upsert_content_type, compose_smtp_payload — fns pures
  de composition du payload SMTP, désormais testables isolément.
- discovery.rs (99 LOC): test_smtp_port, find_smtp_port, expect_code, ehlo_hostname
  — logique de découverte de port SMTP et helpers protocole.
- mod.rs (538 LOC): reste (send_via_relay, send_via_mx, send_email_content*).

Purge dead code: suppression de 4 fns préfixées _ jamais appelées (~148 LOC):
  _parse_email_content, _process_dkim_signature,
  _format_email_address, _validate_email_content
+ import HashMap devenu inutile.

Bénéfices clean code:
- mod.rs 918 → 538 LOC (−41%)
- Séparation des préoccupations: pure fns vs. I/O vs. transport
- Dead code retiré (les 4 fns étaient orphelines depuis longtemps)
- Constants pub(crate) pour usage par sous-modules
