+++ src/smtp_client/mod.rs
@@ -47,253 +47,21 @@ use std::fs::File;
 use std::io::BufReader;
 use uuid::Uuid;
 use webpki_roots::TLS_SERVER_ROOTS;
-const SMTP_PORTS: [u16; 3] = [25, 587, 465];
-const CONNECTION_TIMEOUT: Duration = Duration::from_secs(3);
+pub(crate) const SMTP_PORTS: [u16; 3] = [25, 587, 465];
+pub(crate) const CONNECTION_TIMEOUT: Duration = Duration::from_secs(3);
 use crate::entities::Email;
-use std::collections::HashMap;
 
 enum StreamType {
     Plain(TcpStream),
     Tls(TlsStream<TcpStream>),
 }
 
-fn normalize_crlf(input: &str) -> String {
-    input
-        .replace("\r\n", "\n")
-        .replace('\r', "\n")
-        .replace('\n', "\r\n")
-}
 
-fn strip_tags_simple(html: &str) -> String {
-    let mut out = String::with_capacity(html.len());
-    let mut in_tag = false;
-    for c in html.chars() {
-        match c {
-            '<' => in_tag = true,
-            '>' => in_tag = false,
-            _ if !in_tag => out.push(c),
-            _ => {}
-        }
-    }
-    out.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn body_looks_like_html(body: &str) -> bool {
-    let lower = body.to_ascii_lowercase();
-    lower.contains("<html")
-        || lower.contains("<body")
-        || lower.contains("<p")
-        || lower.contains("<div")
-        || lower.contains("<br")
-        || lower.contains("<table")
-        || lower.contains("<span")
-        || lower.contains("</")
-}
-
-fn ensure_html_document(raw: &str) -> String {
-    let trimmed = raw.trim();
-    let lower = trimmed.to_ascii_lowercase();
-    if lower.contains("<html") {
-        return trimmed.to_string();
-    }
-    format!("<!DOCTYPE html><html><body>{}</body></html>", trimmed)
-}
-
-fn upsert_content_type(headers: &mut Vec<(String, String)>, value: String) {
-    if let Some((_, existing)) = headers
-        .iter_mut()
-        .find(|(k, _)| k.eq_ignore_ascii_case("content-type"))
-    {
-        *existing = value;
-    } else {
-        headers.push(("Content-Type".to_string(), value));
-    }
-}
-
-fn compose_smtp_payload(email: &Email) -> String {
-    let mut headers = email.headers.clone();
-
-    if !headers.iter().any(|(k, _)| k.eq_ignore_ascii_case("date")) {
-        headers.push(("Date".to_string(), Utc::now().to_rfc2822()));
-    }
-    if !headers
-        .iter()
-        .any(|(k, _)| k.eq_ignore_ascii_case("message-id"))
-    {
-        headers.push((
-            "Message-ID".to_string(),
-            format!("<{}@misfits.ai>", Uuid::new_v4()),
-        ));
-    }
-
-    let has_dkim_signature = headers
-        .iter()
-        .any(|(k, _)| k.eq_ignore_ascii_case("dkim-signature"));
-
-    if has_dkim_signature {
-        let mut email_content = String::new();
-
-        let has_from = headers.iter().any(|(k, _)| k.eq_ignore_ascii_case("from"));
-        let has_to = headers.iter().any(|(k, _)| k.eq_ignore_ascii_case("to"));
-        let has_subject = headers.iter().any(|(k, _)| k.eq_ignore_ascii_case("subject"));
-
-        if !has_from {
-            email_content.push_str(&format!("From: {}\r\n", email.from));
-        }
-        if !has_to {
-            email_content.push_str(&format!("To: {}\r\n", email.to));
-        }
-        if !has_subject {
-            email_content.push_str(&format!("Subject: {}\r\n", email.subject));
-        }
-
-        for (key, value) in &headers {
-            email_content.push_str(&format!("{}: {}\r\n", key, value));
-        }
-
-        email_content.push_str("\r\n");
-        email_content.push_str(&normalize_crlf(&email.body));
-        return email_content;
-    }
-
-    headers.retain(|(k, _)| {
-        !(k.eq_ignore_ascii_case("from")
-            || k.eq_ignore_ascii_case("to")
-            || k.eq_ignore_ascii_case("subject"))
-    });
-
-    let body = email.body.as_str();
-
-    let has_multipart = headers.iter().any(|(k, v)| {
-        k.eq_ignore_ascii_case("content-type") && v.to_ascii_lowercase().contains("multipart/")
-    });
-
-    let payload_body = if has_multipart {
-        normalize_crlf(body)
-    } else if body_looks_like_html(body) {
-        let html = normalize_crlf(&ensure_html_document(body));
-        let mut plain = strip_tags_simple(&html);
-        if plain.trim().is_empty() {
-            plain = body.to_string();
-        }
-        plain = normalize_crlf(plain.trim());
-
-        let boundary = format!("misfits-alt-{}", Uuid::new_v4().simple());
-        upsert_content_type(
-            &mut headers,
-            format!("multipart/alternative; boundary=\"{}\"", boundary),
-        );
-
-        format!(
-            "--{b}\r\nContent-Type: text/plain; charset=utf-8\r\nContent-Transfer-Encoding: 8bit\r\n\r\n{plain}\r\n--{b}\r\nContent-Type: text/html; charset=utf-8\r\nContent-Transfer-Encoding: 8bit\r\n\r\n{html}\r\n--{b}--\r\n",
-            b = boundary,
-            plain = plain,
-            html = html,
-        )
-    } else {
-        upsert_content_type(&mut headers, "text/plain; charset=utf-8".to_string());
-        normalize_crlf(body)
-    };
-
-    let mut email_content = format!(
-        "From: {}\r\nTo: {}\r\nSubject: {}\r\n",
-        email.from, email.to, email.subject
-    );
-    for (key, value) in &headers {
-        email_content.push_str(&format!("{}: {}\r\n", key, value));
-    }
-    email_content.push_str(&format!("\r\n{}", payload_body));
-    email_content
-}
-
-async fn test_smtp_port(host: &str, port: u16) -> bool {
-    match timeout(CONNECTION_TIMEOUT, TcpStream::connect((host, port))).await {
-        Ok(Ok(_)) => true,
-        _ => false,
-    }
-}
-
-async fn find_smtp_port(host: &str) -> Option<u16> {
-    for &port in &SMTP_PORTS {
-        if test_smtp_port(host, port).await {
-            return Some(port);
-        }
-    }
-    None
-}
-
-async fn expect_code<T: AsyncReadExt + Unpin>(
-    stream: &mut T,
-    expected: &str,
-) -> std::io::Result<()> {
-    let mut response = [0; 1024];
-    let mut acc = String::new();
-
-    for _ in 0..16 {
-        let n = stream.read(&mut response).await?;
-        if n == 0 {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::UnexpectedEof,
-                format!("Connection closed while waiting for SMTP {}", expected),
-            ));
-        }
-
-        let chunk = String::from_utf8_lossy(&response[..n]);
-        acc.push_str(&chunk);
-        println!("Received response: {}", chunk);
-
-        for line in acc.split("\r\n").filter(|l| !l.is_empty()) {
-            if line.len() < 4 {
-                continue;
-            }
-            let prefix = &line[..3];
-            let sep = line.as_bytes()[3] as char;
-            let is_code = prefix.chars().all(|c| c.is_ascii_digit());
-            if !is_code {
-                continue;
-            }
-
-            if prefix == expected && sep == ' ' {
-                return Ok(());
-            }
-
-            if sep == ' ' && prefix != expected {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    format!("Unexpected response: {}", acc),
-                ));
-            }
-        }
-    }
-
-    Err(std::io::Error::new(
-        std::io::ErrorKind::TimedOut,
-        format!("Timed out waiting for SMTP {} response: {}", expected, acc),
-    ))
-}
-fn ehlo_hostname() -> String {
-    if let Ok(value) = env::var("SMTP_HOSTNAME") {
-        let v = value.trim().trim_end_matches('.');
-        if !v.is_empty() {
-            return v.to_string();
-        }
-    }
-
-    if let Ok(value) = env::var("DOMAIN_NAME") {
-        let d = value.trim().trim_end_matches('.');
-        if !d.is_empty() {
-            if d.eq_ignore_ascii_case("misfits.ai") {
-                return "mail.misfits.ai".to_string();
-            }
-            if d.starts_with("mail.") {
-                return d.to_string();
-            }
-            return format!("mail.{}", d);
-        }
-    }
-
-    "mail.misfits.ai".to_string()
-}
+// Sous-modules extraits pour clean code (refactor architecte).
+mod body_utils;
+mod discovery;
+use body_utils::compose_smtp_payload;
+use discovery::{find_smtp_port, expect_code, ehlo_hostname};
 
 pub async fn send_outgoing_email(email: &Email) -> std::io::Result<()> {
     if let Ok(relay_host) = env::var("SMTP_RELAY_HOST") {
@@ -768,151 +536,3 @@ async fn send_email_content_inner<T: AsyncWriteExt + AsyncReadExt + Unpin>(
     return Ok(());
 }
 
-fn _parse_email_content(content: &str) -> (HashMap<String, String>, String) {
-    let mut headers: HashMap<String, String> = HashMap::new();
-    let mut lines = content.lines().peekable();
-    let mut body = String::new();
-    let mut in_body = false;
-
-    while let Some(line) = lines.next() {
-        if in_body {
-            body.push_str(line);
-            body.push_str("\r\n");
-        } else if line.is_empty() {
-            in_body = true;
-        } else {
-            // New header or continuation
-            if let Some(index) = line.find(':') {
-                let (key, mut value) = line.split_at(index);
-                let header_name = key.trim().to_string();
-                value = value[1..].trim();
-
-                if header_name == "DKIM-Signature" {
-                    let mut full_signature = value.to_string();
-                    let mut b_tag_content = String::new();
-
-                    while let Some(next_line) = lines.next() {
-                        let trimmed = next_line.trim();
-                        if trimmed.is_empty() {
-                            break;
-                        }
-                        if trimmed.starts_with("b=") {
-                            b_tag_content.push_str(&trimmed[2..]); // Start capturing from after "b="
-                            while !b_tag_content.ends_with('=') {
-                                if let Some(next_b_line) = lines.next() {
-                                    b_tag_content.push_str(next_b_line.trim());
-                                } else {
-                                    break; // End of input
-                                }
-                            }
-                            // Add the captured b tag content to full_signature
-                            full_signature.push_str("b=");
-                            full_signature.push_str(&b_tag_content);
-                            break; // We've captured the entire b tag, so we can stop
-                        } else {
-                            full_signature.push(' ');
-                            full_signature.push_str(trimmed);
-                        }
-                    }
-                    println!("Full DKIM-Signature: {}", full_signature);
-                    let processed_signature = _process_dkim_signature(&full_signature);
-                    eprintln!("Processed DKIM-Signature: {}", processed_signature);
-                    headers.insert(header_name, processed_signature);
-                } else {
-                    // Handle other headers
-                    let mut full_value = value.to_string();
-                    while let Some(next_line) = lines.peek() {
-                        if next_line.starts_with(char::is_whitespace) {
-                            full_value.push(' ');
-                            full_value.push_str(next_line.trim());
-                            lines.next(); // consume the peeked line
-                        } else {
-                            break;
-                        }
-                    }
-                    headers.insert(header_name.clone(), full_value.clone());
-                    eprintln!("Inserted header '{}': {}", header_name, full_value);
-                }
-            }
-        }
-    }
-
-    // Process specific headers
-    if let Some(from) = headers.get_mut("From") {
-        let formatted = _format_email_address(from);
-        *from = formatted.clone();
-        eprintln!("Formatted 'From' header: {}", formatted);
-    }
-    if let Some(to) = headers.get_mut("To") {
-        let formatted = _format_email_address(to);
-        *to = formatted.clone();
-        eprintln!("Formatted 'To' header: {}", formatted);
-    }
-    for (key, value) in &headers {
-        eprintln!("Header '{}': {}", key, value);
-    }
-
-    (headers, body)
-}
-
-fn _process_dkim_signature(signature: &str) -> String {
-    // RFC 6376
-    let dkim_tags = [
-        "v", "a", "b", "bh", "c", "d", "h", "i", "l", "q", "s", "t", "x", "z",
-    ];
-
-    let parts: Vec<&str> = signature.split(';').collect();
-    let processed_parts: Vec<String> = parts
-        .iter()
-        .take_while(|&&part| {
-            let trimmed = part.trim();
-            dkim_tags.iter().any(|&tag| {
-                trimmed.starts_with(tag) && trimmed[tag.len()..].trim_start().starts_with('=')
-            })
-        })
-        .map(|&part| {
-            let trimmed = part.trim();
-            if trimmed.starts_with("b=") {
-                // Remove line breaks in 'b' tag value as per DKIM spec
-                let b_value = trimmed.splitn(2, '=').nth(1).unwrap_or("");
-                format!(
-                    "b={}",
-                    b_value
-                        .lines()
-                        .map(|line| line.trim())
-                        .collect::<Vec<_>>()
-                        .join("")
-                )
-            } else {
-                trimmed.to_string()
-            }
-        })
-        .collect();
-
-    processed_parts.join("; ")
-}
-
-fn _format_email_address(addr: &str) -> String {
-    if !addr.contains('<') && !addr.contains('>') {
-        format!("<{}>", addr.trim())
-    } else {
-        addr.to_string()
-    }
-}
-
-fn _validate_email_content(content: &str) -> Result<(), String> {
-    let lines: Vec<&str> = content.lines().collect();
-    if !lines[0].starts_with("From: <") || !lines[0].ends_with(">") {
-        return Err("Invalid From header".to_string());
-    }
-    if !lines[1].starts_with("To: <") || !lines[1].ends_with(">") {
-        return Err("Invalid To header".to_string());
-    }
-    if !lines[2].starts_with("Subject: ") {
-        return Err("Invalid Subject header".to_string());
-    }
-    if lines[3] != "" {
-        return Err("Missing blank line after headers".to_string());
-    }
-    Ok(())
-}
